Add ability to specify uid, gid and group name
In order to facilitate coordination between systems, it can be useful to
be able to specify the uid and gid of the created user. The application
of this most readily at hand is synced folders in an Ansible project
with a VirtualBox provider. When the ids of the host user and the owner
of the files on the guest correspond, permissions issues can be
mitigated.

Contributes to https://github.com/dheles/ansible-role-sufia#7